Feat: removed unnecessary code
